creating integration and testing with ws-raspay

+++ src/main/java/com/client/ws/rasmooplus/dto/wsraspay/OrderDto.java
@@ -8,14 +8,14 @@ public class OrderDto {
 
     private String customerId;
 
-    private Long discount;
+    private BigDecimal discount;
 
     private String productAcronym;
 
     public OrderDto() {
     }
 
-    public OrderDto(String id, String customerId, Long discount, String productAcronym) {
+    public OrderDto(String id, String customerId, BigDecimal discount, String productAcronym) {
         this.id = id;
         this.customerId = customerId;
         this.discount = discount;
@@ -38,11 +38,11 @@ public class OrderDto {
         this.customerId = customerId;
     }
 
-    public Long getDiscount() {
+    public BigDecimal getDiscount() {
         return discount;
     }
 
-    public void setDiscount(Long discount) {
+    public void setDiscount(BigDecimal discount) {
         this.discount = discount;
     }
